requirements.txt
streamlit
google-generativeai
python-pptx 